+++ deployment/fabfile.py
@@ -1,7 +1,7 @@
 from fabric import task
 
 @task
-def deploy_next(c):
+def deploy_pb(c):
     '''Deploys the next version of the Peerbook server to the next service'''
     next = c.run('basename $(readlink /opt/peerbook/next)').stdout
     try:
@@ -13,6 +13,14 @@ def deploy_next(c):
     c.put('peerbook', '/opt/peerbook/next')
     c.sudo(f'supervisorctl restart pb-{next}')
     c.sudo("nginx -s reload")
+
+@task
+def deploy_host(c):
+    '''wIP: Deploys the host archive to the next service'''
+    with c.cd('./host_archive'):
+        c.local('tar -czf ../archive.tar.gz .')
+    c.put('archive.tar.gz', '.')
+    c.run('tar -xzf archive.tar.gz -C /')
 
 @task
 def switch(c):